requirements.txt
beautifulsoup4==4.9.3
Pillow==8.2.0
gitpython==3.1.14
requests==2.25.1